feat: 修正信号掩码 ABI 编码
- 使用 SignalNum 统一维护 Linux signal 编号
- 使用 SignalBit 区分 signal 编码位图，按 1 << (signum - 1) 构造
- 将内核和用户态 sigset_t 路径统一为 64 位处理
- 兼容 4 字节 sigsetsize 输入，并裁剪未支持的信号位
- 修正 sigprocmask、sigsuspend、sigtimedwait、sigreturn 的 mask 读写逻辑
- 更新 signal TODO 文档，记录后续 signal 机制待完善项

+++ user/src/syscall.rs
@@ -443,7 +443,7 @@ pub fn sys_sigaction(
 }
 
 /// New rt_sigprocmask ABI wrapper: (how, set, oset, sigsetsize)
-pub fn sys_rt_sigprocmask(how: i32, set: *const u32, oset: *mut u32, sigsetsize: usize) -> isize {
+pub fn sys_rt_sigprocmask(how: i32, set: *const u64, oset: *mut u64, sigsetsize: usize) -> isize {
     syscall6(
         SYSCALL_SIGPROCMASK,
         [how as usize, set as usize, oset as usize, sigsetsize, 0, 0],
@@ -451,10 +451,10 @@ pub fn sys_rt_sigprocmask(how: i32, set: *const u32, oset: *mut u32, sigsetsize:
 }
 
 /// Compatibility helper that sets the mask to `mask` (SIG_SETMASK semantics).
-pub fn sys_sigprocmask(mask: u32) -> isize {
+pub fn sys_sigprocmask(mask: u64) -> isize {
     // SIG_SETMASK == 2
-    let size = core::mem::size_of::<u32>();
-    sys_rt_sigprocmask(2, &mask as *const u32, core::ptr::null_mut(), size)
+    let size = core::mem::size_of::<u64>();
+    sys_rt_sigprocmask(2, &mask as *const u64, core::ptr::null_mut(), size)
 }
 
 pub fn sys_sigreturn() -> isize {
